[utils] add an option to select default user for rabbitmq and db
This change introduces an option to select the default user for RabbitMQ
and MariaDB template function:
* `ini_sections.default_transport_url`
* `utils.db_url` with sub-functions

Currently, these functions select the default user the following way:
* RabbitMQ selects the 'default' user.
* MariaDB selects the `first` defined database and then chooses the key
from users with the same name as the database.

After this change, `defaultUser` option (`.Values.rabbitmq.defaultUser`
or `.Values.mariadb.defaultUser`) could be used to select another user
when required.